requirements.txt
python-dotenv
requests
schedule
apscheduler
google-auth
google-api-python-client
openai
pytz>=2023.3

